+++ Lesson-6/assignment/payroll/contracts/Payroll.sol
@@ -76,6 +76,13 @@ contract Payroll is Ownable {
         totalSalary = totalSalary.sub(employee.salary);
         delete employees[employeeId];
         totalEmployee = totalEmployee.sub(1);
+        for (uint i = 0; i < employeeList.length; i++) {
+            if (employeeList[i] == employeeId) {
+                delete employeeList[i];
+                employeeList[i] = employeeList[employeeList.length - 1];
+                employeeList.length -= 1;
+            }
+        }
         RemoveEmployee(employeeId);
     }
