+++ Lesson-6/assignment/payroll/src/components/Employee.js
@@ -14,19 +14,20 @@ class Employer extends Component {
   }
 
   checkEmployee = () => {
-    const { payroll, employee, web3 } = this.props;
-    payroll.employees.call(employee, {
-      from: employee,
+    const { payroll, account, web3 } = this.props;
+    payroll.employees.call(account, {
+      from: account,
       gas: 1000000
     }).then((result) =>{
       console.log(result)
       this.setState({
         salary: web3.fromWei(result[1].toNumber()),
-        lastPaidDate: new Date(result[2].toNumber() * 1000)
+        lastPaidDate: new Date(result[2].toNumber() * 1000).toString()
       });
+      console.log(this.salary);
     });
-
-    web3.eth.getBalance(employee, (err, result) => {
+    
+    web3.eth.getBalance(account, (err, result) => {
       this.setState({
         balance: web3.fromWei(result.toNumber())
       });
@@ -34,9 +35,9 @@ class Employer extends Component {
   }
 
   getPaid = () => {
-    const { payroll, employee} = this.props;
+    const { payroll, account} = this.props;
     payroll.getPaid({
-      from: employee,
+      from: account,
       gas: 1000000
     }).then((result) => {
       alert('You have been paid');
